refactor: move theme dependent props to hook
