+++ app/src/pages/report.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+//import { GlobalContext } from '../contexts/global';
+import type { NextPage } from 'next';
+//import { useRouter } from 'next/router';
+//import { Layout } from '../components';
+
+const ReportPage: NextPage = () => {
+  return <div>{`page`}</div>;
+};
+
+export default ReportPage;
